[add] <tauri> api_handler
