+++ src-tauri/tauri.conf.json
@@ -1,7 +1,7 @@
 {
   "package": {
     "productName": "Smart Dental",
-    "version": "v0.0.2"
+    "version": "0.0.2"
   },
   "build": {
     "distDir": "../dist",
